refactor: user - 디렉토리 재배치, userService에서 중복 코드 메서드로 추출#63
