+++ user-service/src/main/java/com/oringmaryho/business/userservice/config/pageable/PageableConfig.java
@@ -10,6 +10,8 @@ import org.springframework.data.web.config.EnableSpringDataWebSupport;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
+import com.oringmaryho.business.userservice.utils.SortConstants;
+
 @Configuration
 @EnableSpringDataWebSupport(pageSerializationMode = EnableSpringDataWebSupport.PageSerializationMode.VIA_DTO)
 public class PageableConfig implements WebMvcConfigurer {
