+++ user-service/src/test/java/com/oringmaryho/business/userservice/application/UserServiceTest.java
@@ -12,6 +12,7 @@ import org.mockito.MockitoAnnotations;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
 import com.oringmaryho.business.userservice.application.dto.request.UserSignUpRequestServiceDto;
+import com.oringmaryho.business.userservice.application.service.UserService;
 import com.oringmaryho.business.userservice.domain.User;
 import com.oringmaryho.business.userservice.domain.repository.UserRepository;
